voicecallmanager: prevent multiple rtcom logging
If we hang up the call ourselves, then will call hangup(), log the event, and
then later on statusChanged() is called with the disconnected status, and then
we'd log it again.  Let's just keep track of whether we emitted a disconnected
call state change or not as a private class variable.